+++ brickrail-gui/ble/ble_controller.gd
@@ -43,7 +43,7 @@ func _on_hub_removing(hubname):
 
 func _on_hub_state_changed(_hub=null):
 	if $BLECommunicator.connected:
-		if not are_hubs_ready() and LayoutInfo.control_devices:
+		if not are_hubs_ready() and LayoutInfo.control_devices==LayoutInfo.CONTROL_ALL:
 			LayoutInfo.emergency_stop()
 		var status = get_hub_status()
 		if len(status)>0:
@@ -54,8 +54,8 @@ func _on_hub_state_changed(_hub=null):
 		
 		hub_control_enabled = (len(status) == 0)
 	else:
-		if LayoutInfo.control_devices:
-			LayoutInfo.set_control_devices(false)
+		if LayoutInfo.control_devices>0:
+			LayoutInfo.set_control_devices(LayoutInfo.CONTROL_OFF)
 			LayoutInfo.stop_all_trains()
 		if $BLECommunicator.busy:
 			GuiApi.status_process("[BLE Server] "+$BLECommunicator.status+"...")