+++ brickrail-gui/devices/train/ble_train.gd
@@ -91,7 +91,7 @@ func download_route(p_route):
 		hub.rpc("set_route_leg", data)
 
 func _on_route_intention_changed(leg_index, intention):
-	if not LayoutInfo.control_devices:
+	if not LayoutInfo.control_devices==LayoutInfo.CONTROL_ALL:
 		return
 	hub.rpc("set_leg_intention", [leg_index, intention_to_enum[intention]])